preproc: added soaPartitioning function
Added the "SOA partitioning" function. Allocates aSOA/bSOA based on the linear regression of aVOC/bVOC vs tSOA. Some problems right now where the noise is overwhelming the signal - need to see if this persists in the actual dataset or if it is an artifact of the "testdata" averaging/simplification. Might also be fixed just by having more time steps?