new_gemmach: updated landuse and added QC import from separate file
2.	Still have some problems with “stability mixing chemistry”
a.	Problem is in the landuse function
3.	Found a couple of small bugs in the “fractions” function
a.	As written, this function was adding the total area of intersection for each category to some total area then dividing (each time) by the total polygon area. Made two changes:
i.	Moved the final divide out of the polygon – we want the fraction in each polygon, so it doesn’t make sense to divide every time. This matches the GEOSCHEM format.
ii.	Since for our values each grid cell could have multiple fractions, we need to multiply by that fraction – which is stored as c.vf
b.	This was still giving a problem where the output of a 100% ocean cell was being recorded as 0.25, e.g. sum of fractions didn’t equal 1.
i.	So, either the polygons aren’t overlapping or it is the wrong area.
ii.	Looks like they aren’t overlapping
1.	SB to fix.
4.	GEMSeinfeldLandUse function was returning an error – need to subtract 1 from output value. Updated.
a.	Still producing an error – shape was coming from 3D variable, landuse is 2d. Switched to PBLH. Now it works!