vargrid.go, gemmach.go testing gemmach outputs
Tried to use GEMMACH preproc outputs through the rest of the code and had problems reprojecting (we were using rlat, rlon). So here, we have converted all GEMMACH preproc inputs to LCC, and confirmed that it worked. Major changes are switching geophy values from delta_rlat/rlon to delta_x/y, and similarly the reading of x and y values from rlat/rlon to x/y coordinates. Also some testing in the vargrid file to get our outputs to work - we switched the "default" grid that is created to be 4x4 (16 x-y cells) rather than 2x2. Next step is testing this configuration on the full dataset.